+++ Exercism/c/grains/src/grains.h
@@ -1,0 +1,8 @@
+#ifndef GRAINS_H
+#define GRAINS_H
+
+unsigned long square(int);
+
+unsigned long total();
+
+#endif